+++ core-api/src/main/java/com/optimizely/ab/event/internal/EventBuilderV2.java
@@ -1,5 +1,6 @@
-/*
- *    Copyright 2017, Optimizely
+/**
+ *
+ *    Copyright 2016, Optimizely and contributors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,7 +16,6 @@
  */
 package com.optimizely.ab.event.internal;
 
-import com.optimizely.ab.BuildConfig;
 import com.optimizely.ab.annotations.VisibleForTesting;
 import com.optimizely.ab.bucketing.Bucketer;
 import com.optimizely.ab.config.Attribute;
@@ -63,7 +63,7 @@ public class EventBuilderV2 extends EventBuilder {
     private Serializer serializer;
 
     public EventBuilderV2() {
-        this(ClientEngine.JAVA_SDK, BuildConfig.VERSION);
+        this(ClientEngine.JAVA_SDK, BuildVersionInfo.VERSION);
     }
 
     public EventBuilderV2(ClientEngine clientEngine, String clientVersion) {