+++ core-api/src/main/java/com/optimizely/ab/notification/NotificationBroadcaster.java
@@ -1,5 +1,6 @@
-/*
- *    Copyright 2017, Optimizely
+/**
+ *
+ *    Copyright 2016, Optimizely and contributors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -18,6 +19,7 @@ package com.optimizely.ab.notification;
 import com.optimizely.ab.annotations.VisibleForTesting;
 import com.optimizely.ab.config.Experiment;
 import com.optimizely.ab.config.Variation;
+import com.optimizely.ab.event.LogEvent;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,6 +27,7 @@ import org.slf4j.LoggerFactory;
 import java.util.HashSet;
 import java.util.Map;
 
+import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 
 /**
@@ -73,6 +76,25 @@ public class NotificationBroadcaster {
     public void clearListeners() {
         listeners.clear();
         logger.debug("Notification listeners were cleared");
+    }
+
+    /**
+     * Notify listeners that an Optimizely event has been tracked.
+     *
+     * @param eventKey the key of the tracked event
+     * @param userId the ID of the user
+     * @param attributes a map of attributes about the event
+     * @param eventValue an integer to be aggregated for the event
+     * @param logEvent the log event sent to the event dispatcher
+     */
+    public void broadcastEventTracked(@Nonnull String eventKey,
+                                      @Nonnull String userId,
+                                      @Nonnull Map<String, String> attributes,
+                                      @CheckForNull Long eventValue,
+                                      @Nonnull LogEvent logEvent) {
+        for (final NotificationListener iterListener : listeners) {
+            iterListener.onEventTracked(eventKey, userId, attributes, eventValue, logEvent);
+        }
     }
 
     /**