+++ core-api/src/test/java/com/optimizely/ab/OptimizelyTestV3.java
@@ -1,5 +1,6 @@
-/*
- *    Copyright 2017, Optimizely
+/**
+ *
+ *    Copyright 2016, Optimizely and contributors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -1115,11 +1116,11 @@ public class OptimizelyTestV3 {
     }
 
     /**
-     * Verify that {@link Optimizely#getVariableFloat(String, String, Map, boolean)} returns a float live variable
+     * Verify that {@link Optimizely#getVariableDouble(String, String, Map, boolean)} returns a double live variable
      * value when an proper variable key is provided and dispatches an impression when activateExperiment is true.
      */
     @Test
-    public void getVariableFloat() throws Exception {
+    public void getVariableDouble() throws Exception {
         String datafile = validConfigJsonV3();
         ProjectConfig projectConfig = validProjectConfigV3();
 
@@ -1134,9 +1135,9 @@ public class OptimizelyTestV3 {
             .withBucketing(mockBucketer)
             .build();
 
-        assertThat(optimizely.getVariableFloat("float_variable", "userId",
+        assertThat(optimizely.getVariableDouble("double_variable", "userId",
                                                Collections.singletonMap("browser_type", "chrome"), true),
-                   is(5.3f));
+                   is(5.3));
         verify(mockEventHandler).dispatchEvent(any(LogEvent.class));
     }
 
@@ -1506,6 +1507,19 @@ public class OptimizelyTestV3 {
         optimizely.getVariableString("string_variable", userId, attributes, activateExperiment);
         verify(listener, times(2))
                 .onExperimentActivated(activatedExperiment, userId, attributes, actualVariation);
+
+        // Check if listener is notified after an event is tracked
+        EventType eventType = projectConfig.getEventTypes().get(0);
+        String eventKey = eventType.getKey();
+
+        when(mockEventBuilder.createConversionEvent(eq(projectConfig), eq(mockBucketer), eq(userId),
+                eq(eventType.getId()), eq(eventKey),
+                anyMapOf(String.class, String.class)))
+                .thenReturn(logEventToDispatch);
+
+        optimizely.track(eventKey, userId, attributes);
+        verify(listener, times(1))
+                .onEventTracked(eventKey, userId, attributes, null, logEventToDispatch);
     }
 
     /**
@@ -1560,6 +1574,19 @@ public class OptimizelyTestV3 {
         optimizely.getVariableString("string_variable", userId, attributes, activateExperiment);
         verify(listener, never())
                 .onExperimentActivated(activatedExperiment, userId, attributes, actualVariation);
+
+        // Check if listener is notified after an event is tracked
+        EventType eventType = projectConfig.getEventTypes().get(0);
+        String eventKey = eventType.getKey();
+
+        when(mockEventBuilder.createConversionEvent(eq(projectConfig), eq(mockBucketer), eq(userId),
+                eq(eventType.getId()), eq(eventKey),
+                anyMapOf(String.class, String.class)))
+                .thenReturn(logEventToDispatch);
+
+        optimizely.track(eventKey, userId, attributes);
+        verify(listener, never())
+                .onEventTracked(eventKey, userId, attributes, null, logEventToDispatch);
     }
 
     /**
@@ -1613,6 +1640,19 @@ public class OptimizelyTestV3 {
         optimizely.getVariableString("string_variable", userId, attributes, activateExperiment);
         verify(listener, never())
                 .onExperimentActivated(activatedExperiment, userId, attributes, actualVariation);
+
+        // Check if listener is notified after a event is tracked
+        EventType eventType = projectConfig.getEventTypes().get(0);
+        String eventKey = eventType.getKey();
+
+        when(mockEventBuilder.createConversionEvent(eq(projectConfig), eq(mockBucketer), eq(userId),
+                eq(eventType.getId()), eq(eventKey),
+                anyMapOf(String.class, String.class)))
+                .thenReturn(logEventToDispatch);
+
+        optimizely.track(eventKey, userId, attributes);
+        verify(listener, never())
+                .onEventTracked(eventKey, userId, attributes, null, logEventToDispatch);
     }
 
     //======== Helper methods ========//