+++ core-api/src/test/java/com/optimizely/ab/notification/NotificationBroadcasterTest.java
@@ -1,5 +1,6 @@
-/*
- *    Copyright 2017, Optimizely
+/**
+ *
+ *    Copyright 2016, Optimizely and contributors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -17,6 +18,7 @@ package com.optimizely.ab.notification;
 
 import com.optimizely.ab.config.Experiment;
 import com.optimizely.ab.config.Variation;
+import com.optimizely.ab.event.LogEvent;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -101,6 +103,26 @@ public class NotificationBroadcasterTest {
         notificationBroadcaster.clearListeners();
         assertEquals("clearListeners did not remove all listeners that were added",
                 0, notificationBroadcaster.listeners.size());
+    }
+
+    /**
+     * Verify that {@link NotificationBroadcaster#broadcastEventTracked(String, String, Map, Long, LogEvent)}
+     * notifies all listeners.
+     */
+    @Test
+    public void broadcastEventTracked() throws Exception {
+        notificationBroadcaster.addListener(listener);
+        notificationBroadcaster.addListener(listener2);
+
+        String eventKey = "event1";
+        String userId = "user1";
+        Map<String, String> attributes = Collections.emptyMap();
+        Long eventValue = 0L;
+        LogEvent logEvent = mock(LogEvent.class);
+        notificationBroadcaster.broadcastEventTracked(
+                eventKey, userId, attributes, eventValue, logEvent);
+        verify(listener).onEventTracked(eventKey, userId, attributes, eventValue, logEvent);
+        verify(listener2).onEventTracked(eventKey, userId, attributes, eventValue, logEvent);
     }
 
     /**